+++ config/storybook/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @marigold/storybook-config
 
+## 1.1.50
+
+### Patch Changes
+
+- Updated dependencies [65608b4]
+- Updated dependencies [caefbe4]
+- Updated dependencies [2d9917f]
+- Updated dependencies [6f8e3a2]
+- Updated dependencies [7ea3838]
+- Updated dependencies [6687af7]
+- Updated dependencies [2babc0b]
+- Updated dependencies [8c539d3]
+- Updated dependencies [f18c8aa]
+- Updated dependencies [d5386e4]
+- Updated dependencies [5798b4d]
+- Updated dependencies [5c029ec]
+- Updated dependencies [2169b6f]
+- Updated dependencies [bfd2843]
+- Updated dependencies [0e77996]
+- Updated dependencies [b8cd92a]
+- Updated dependencies [45fb3c4]
+- Updated dependencies [f6a132c]
+- Updated dependencies [956982a]
+- Updated dependencies [df04623]
+- Updated dependencies [f3f0ea0]
+  - @marigold/components@10.0.0
+  - @marigold/theme-b2b@28.0.0
+  - @marigold/theme-core@27.0.0
+
 ## 1.1.49
 
 ### Patch Changes